add ollama versioning

+++ DanaCopilot.Domain/Entities/DocumentChunk.cs
@@ -7,7 +7,7 @@ namespace DanaCopilot.Domain
     public class DocumentChunk
     {
         public long Id { get; set; }
-        public long ChunkIndex { get; set; }
+        public int ChunkIndex { get; set; }
         public long DocumentId { get; set; }
         public string Content { get; set; }
         public string ContentHash { get; set; }
